Tratamento de erros/ Modificaç

+++ src/modules/users/contexts/delete/delete.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'shared/database/entities/user.entity';
 import { UserRepository } from 'shared/repositories/user.repository';
+import { EntityNotFoundError } from 'typeorm';
 
 @Injectable()
 export class DeleteUserService {
@@ -10,6 +12,8 @@ export class DeleteUserService {
   ) {}
 
   async remove(id: string) {
+    const user = this.repository.findOneUser(id);
+    if (!user) throw new EntityNotFoundError(User, id);
     await this.repository.deleteUser(id);
   }
 }
